Move webdriver spy mode functionality to internal class WebdriverProtocol
