--- testar/resources/settings/webdriver_spy_custom/Protocol_webdriver_spy_custom.java
@@ -1,214 +1,0 @@
-/**
- * Copyright (c) 2018 - 2022 Open Universiteit - www.ou.nl
- * Copyright (c) 2019 - 2022 Universitat Politecnica de Valencia - www.upv.es
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution.
- * 3. Neither the name of the copyright holder nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- */
-
-import org.testar.monkey.Util;
-import org.testar.monkey.alayer.*;
-import org.testar.monkey.alayer.actions.AnnotatingActionCompiler;
-import org.testar.monkey.alayer.actions.StdActionCompiler;
-import org.testar.monkey.alayer.actions.WdHistoryBackAction;
-import org.testar.monkey.alayer.devices.KBKeys;
-import org.testar.monkey.alayer.exceptions.ActionBuildException;
-import org.testar.monkey.alayer.webdriver.WdElement;
-import org.testar.monkey.alayer.webdriver.WdWidget;
-import org.testar.monkey.alayer.webdriver.enums.WdRoles;
-import org.testar.monkey.alayer.webdriver.enums.WdTags;
-import org.testar.plugin.NativeLinker;
-import org.testar.SutVisualization;
-import org.testar.managers.InputDataManager;
-import org.testar.monkey.Main;
-import org.testar.protocols.WebdriverProtocol;
-
-import java.util.*;
-import static org.testar.monkey.alayer.Tags.Blocked;
-import static org.testar.monkey.alayer.Tags.Enabled;
-import static org.testar.monkey.alayer.webdriver.Constants.scrollArrowSize;
-import static org.testar.monkey.alayer.webdriver.Constants.scrollThick;
-
-/**
- * Customize the clickable css classes using TESTAR Spy mode + Control button over widget. 
- */
-public class Protocol_webdriver_spy_custom extends WebdriverProtocol {
-
-	/**
-	 * This method is used by TESTAR to determine the set of currently available actions.
-	 * You can use the SUT's current state, analyze the widgets and their properties to create
-	 * a set of sensible actions, such as: "Click every Button which is enabled" etc.
-	 * The return value is supposed to be non-null. If the returned set is empty, TESTAR
-	 * will stop generation of the current action and continue with the next one.
-	 *
-	 * @param system the SUT
-	 * @param state  the SUT's current state
-	 * @return a set of actions
-	 */
-	@Override
-	protected Set<Action> deriveActions(SUT system, State state) throws ActionBuildException {
-		// Kill unwanted processes, force SUT to foreground
-		Set<Action> actions = super.deriveActions(system, state);
-		Set<Action> filteredActions = new HashSet<>();
-
-		// create an action compiler, which helps us create actions
-		// such as clicks, drag&drop, typing ...
-		StdActionCompiler ac = new AnnotatingActionCompiler();
-
-		// Check if forced actions are needed to stay within allowed domains
-		Set<Action> forcedActions = detectForcedActions(state, ac);
-
-		// iterate through all widgets
-		for (Widget widget : state) {
-			// only consider enabled and non-tabu widgets
-			if (!widget.get(Enabled, true)) {
-				continue;
-			}
-			// The blackListed widgets are those that have been filtered during the SPY mode with the
-			//CAPS_LOCK + SHIFT + Click clickfilter functionality.
-			if(blackListed(widget)){
-				if(isTypeable(widget)){
-					filteredActions.add(ac.clickTypeInto(widget, InputDataManager.getRandomTextInputData(widget), true));
-				} else {
-					filteredActions.add(ac.leftClickAt(widget));
-				}
-				continue;
-			}
-
-			// slides can happen, even though the widget might be blocked
-			addSlidingActions(actions, ac, scrollArrowSize, scrollThick, widget);
-
-			// If the element is blocked, Testar can't click on or type in the widget
-			if (widget.get(Blocked, false) && !widget.get(WdTags.WebIsShadow, false)) {
-				continue;
-			}
-
-			// type into text boxes
-			if (isAtBrowserCanvas(widget) && isTypeable(widget)) {
-				if(whiteListed(widget) || isUnfiltered(widget)){
-					actions.add(ac.clickTypeInto(widget, InputDataManager.getRandomTextInputData(widget), true));
-				}else{
-					// filtered and not white listed:
-					filteredActions.add(ac.clickTypeInto(widget, InputDataManager.getRandomTextInputData(widget), true));
-				}
-			}
-
-			// left clicks, but ignore links outside domain
-			if (isAtBrowserCanvas(widget) && isClickable(widget)) {
-				if(whiteListed(widget) || isUnfiltered(widget)){
-					if (!isLinkDenied(widget)) {
-						actions.add(ac.leftClickAt(widget));
-					}else{
-						// link denied:
-						filteredActions.add(ac.leftClickAt(widget));
-					}
-				}else{
-					// filtered and not white listed:
-					filteredActions.add(ac.leftClickAt(widget));
-				}
-			}
-		}
-
-		if(actions.isEmpty()) {
-			return new HashSet<>(Collections.singletonList(new WdHistoryBackAction()));
-		}
-
-		// If we have forced actions, prioritize and filter the other ones
-		if (forcedActions != null && forcedActions.size() > 0) {
-			filteredActions = actions;
-			actions = forcedActions;
-		}
-
-		//Showing the grey dots for filtered actions if visualization is on:
-		if(visualizationOn || mode() == Modes.Spy) SutVisualization.visualizeFilteredActions(cv, state, filteredActions);
-
-		return actions;
-	}
-
-	@Override
-	protected boolean isClickable(Widget widget) {
-		Role role = widget.get(Tags.Role, Roles.Widget);
-		if (Role.isOneOf(role, NativeLinker.getNativeClickableRoles())) {
-			// Input type are special...
-			if (role.equals(WdRoles.WdINPUT)) {
-				String type = ((WdWidget) widget).element.type;
-				return WdRoles.clickableInputTypes().contains(type);
-			}
-			return true;
-		}
-
-		WdElement element = ((WdWidget) widget).element;
-		if (element.isClickable) {
-			return true;
-		}
-
-		Set<String> clickSet = new HashSet<>(clickableClasses);
-		clickSet.retainAll(element.cssClasses);
-		return clickSet.size() > 0;
-	}
-
-
-	/**
-	 * Add additional TESTAR keyboard shortcuts in SPY mode to enable the CSS clickable customization of widgets. 
-	 * @param key
-	 */
-	@Override
-	public void keyDown(KBKeys key) {    	
-		super.keyDown(key);        
-		if (mode() == Modes.Spy){ 
-			if (key == KBKeys.VK_RIGHT) {
-				try {
-					// Obtain the widget aimed with the mouse cursor
-					Widget w = Util.widgetFromPoint(latestState, mouse.cursor().x(), mouse.cursor().y());
-					// Add the widget web CSS class property as clickable
-					WdElement element = ((WdWidget) w).element;
-					for(String s : element.cssClasses)
-						if(s!=null && !s.isEmpty())
-							clickableClasses.add(s);
-					// And save the new CSS class property in the test.setting file
-					Util.saveToFile(settings.toFileString(), Main.getTestSettingsFile());
-				} catch(Exception e) {
-					System.out.println("ERROR adding the widget from point: " + "x(" + mouse.cursor().x() + "), y("+ mouse.cursor().y() +")");
-				}
-			}
-
-			if (key == KBKeys.VK_LEFT) {
-				try {
-					// Obtain the widget aimed with the mouse cursor
-					Widget w = Util.widgetFromPoint(latestState, mouse.cursor().x(), mouse.cursor().y());
-					// Remove the widget web CSS class property from all clickables
-					WdElement element = ((WdWidget) w).element;
-					for(String s : element.cssClasses)
-						if(s!=null && !s.isEmpty())
-							clickableClasses.remove(s);
-					// And save the new CSS class property in the test.setting file
-					Util.saveToFile(settings.toFileString(), Main.getTestSettingsFile());
-				} catch(Exception e) {
-					System.out.println("ERROR removing the widget from point: " + "x(" + mouse.cursor().x() + "), y("+ mouse.cursor().y() +")");
-				}
-			}
-		}
-	}
-}
